backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-pymongo==2.3.0
pymongo==3.12.0
requests==2.26.0
python-dotenv==0.19.0
google-cloud-speech==2.11.0
google-generativeai==0.3.0
waitress==2.0.0
python-multipart==0.0.5
werkzeug==2.0.3
uvicorn==0.15.0
fastapi==0.68.1
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
dnspython==2.3.0 
geopy==2.4.1
google-cloud-firestore
google-cloud-storage
pika
pytz